build: add more request metrics to Posthog
Ref: metriport/metriport-internal#136
