+++ packages/api/src/routes/helpers/request-analytics.ts
@@ -1,0 +1,55 @@
+import { Request } from "express";
+import { Product } from "../../domain/product";
+import { EventTypes, analytics } from "../../shared/analytics";
+import { getCxId, getCxIdFromHeaders } from "../util";
+
+const medicalRoutes = ["medical", "fhir"];
+const devicesRoutes = [
+  "activity",
+  "body",
+  "biometrics",
+  "nutrition",
+  "sleep",
+  "user",
+  "connect",
+  "garmin",
+  "withings",
+  "apple",
+  "fitbit",
+  "tenovi",
+];
+
+export const analyzeRoute = ({
+  req,
+  method,
+  url,
+  duration,
+}: {
+  req: Request;
+  method: string;
+  url: string;
+  duration: number;
+}): void => {
+  const reqCxId = getCxId(req);
+  const headerCxId = getCxIdFromHeaders(req);
+  const cxId = reqCxId ?? headerCxId;
+
+  const isMedical = medicalRoutes.some(route => req.baseUrl.includes(route));
+  const isDevices = devicesRoutes.some(route => req.baseUrl.includes(route));
+
+  const distinctId = cxId ?? "not-available";
+  analytics({
+    distinctId,
+    event: EventTypes.query,
+    properties: {
+      method,
+      url,
+      ...(isMedical
+        ? { apiType: Product.medical }
+        : isDevices
+        ? { apiType: Product.devices }
+        : { apiType: "internal" }),
+      duration,
+    },
+  });
+};
