+++ packages/api/src/routes/helpers/request-logger.ts
@@ -1,4 +1,5 @@
 import { NextFunction, Request, Response } from "express";
+import { analyzeRoute } from "./request-analytics";
 
 export const requestLogger = (req: Request, res: Response, next: NextFunction): void => {
   const method = req.method;
@@ -11,6 +12,8 @@ export const requestLogger = (req: Request, res: Response, next: NextFunction): 
     const elapsedHrTime = process.hrtime(startHrTime);
     const elapsedTimeInMs = elapsedHrTime[0] * 1000 + elapsedHrTime[1] / 1e6;
     console.log("..........Done %s %s | %d | %fms", method, url, res.statusCode, elapsedTimeInMs);
+
+    analyzeRoute({ req, method, url, duration: elapsedTimeInMs });
   });
   next();
 };
