+++ packages/api/src/routes/util.ts
@@ -1,7 +1,5 @@
 import { NextFunction, Request, Response } from "express";
-import { Product } from "../domain/product";
 import BadRequestError from "../errors/bad-request";
-import { analytics, EventTypes } from "../shared/analytics";
 import { Config } from "../shared/config";
 import { errorToString } from "../shared/log";
 import { capture } from "../shared/notifications";
@@ -17,7 +15,6 @@ export const asyncHandler =
   ) =>
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     try {
-      analyzeRoute(req);
       await f(req, res, next);
     } catch (err) {
       if (Config.isCloudEnv()) console.error(errorToString(err));
@@ -25,45 +22,6 @@ export const asyncHandler =
       next(err);
     }
   };
-
-export const analyzeRoute = (req: Request): void => {
-  const medicalRoutes = ["medical", "fhir"];
-  const devicesRoutes = ["activity", "body", "biometrics", "nutrition", "sleep", "user"];
-
-  let cxId;
-
-  const reqCxId = getCxId(req);
-  if (reqCxId) cxId = reqCxId;
-
-  const headerCxId = getCxIdFromHeaders(req);
-  if (headerCxId) cxId = headerCxId;
-
-  if (cxId) {
-    const isMedical = medicalRoutes.some(route => req.baseUrl.includes(route));
-    const isDevices = devicesRoutes.some(route => req.baseUrl.includes(route));
-
-    let reqUrl = req.baseUrl;
-    const hasPath = req.route.path.split("/")[1];
-
-    if (hasPath) {
-      reqUrl = reqUrl.concat(req.route.path);
-    }
-
-    analytics({
-      distinctId: cxId,
-      event: EventTypes.query,
-      properties: {
-        method: req.method,
-        url: reqUrl,
-        ...(isMedical
-          ? { apiType: Product.medical }
-          : isDevices
-          ? { apiType: Product.devices }
-          : undefined),
-      },
-    });
-  }
-};
 
 // https://www.rfc-editor.org/rfc/rfc7807 w/ Metriport extension, { name?: string }
 export type HttpResponseBody = { status: number; title: string; detail?: string; name?: string };
